cloud

+++ case 3.2 securitybig/src/main/java/com/security/demo/repository/UserRepository.java
@@ -10,5 +10,6 @@ public interface UserRepository extends JpaRepository<Users,Long> {
 
     Users findByUsername(String username);
 
-    Users findById(long Id);
+    Users findById(Integer Id);
+
 }
